tests/wpt/web-platform-tests/2dcontext/pixel-manipulation/2d.imageData.create2.zero.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.imageData.create2.zero</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.imageData.create2.zero</h1>
<p class="desc">createImageData(sw, sh) throws INDEX_SIZE_ERR if size is zero</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("createImageData(sw, sh) throws INDEX_SIZE_ERR if size is zero");
_addTest(function(canvas, ctx) {

assert_throws("INDEX_SIZE_ERR", function() { ctx.createImageData(10, 0); });
assert_throws("INDEX_SIZE_ERR", function() { ctx.createImageData(0, 10); });
assert_throws("INDEX_SIZE_ERR", function() { ctx.createImageData(0, 0); });
assert_throws("INDEX_SIZE_ERR", function() { ctx.createImageData(0.99, 10); });
assert_throws("INDEX_SIZE_ERR", function() { ctx.createImageData(10, 0.1); });


});
</script>
